sm8250-common: Place checks before setting new ringer mode (alert slider)
The keyEvent is being called 2 times on a single change and consequently
the ringer mode is set twice as well. The audio manager provides it's
own haptic feedback on mode change to RINGER_MODE_NORMAL, so calling it
twice leads to multiple vibrations and the result is a shitty feedback.
This solves that and we now use EFFECT_POP instead of 50ms.